requirements.txt
pillow==7.0.0
face-recognition==1.3.0 
scikit-image==0.16.2 